fixes silo and xeno burrowed to marine ratio issue.

+++ code/modules/mob/living/carbon/xenomorph/hive_datum.dm
@@ -682,13 +682,18 @@ to_chat will check for valid clients itself already so no need to double check f
 			continue
 		boarder.gib()
 		left_behind++
+	var/datum/job/xeno_job = SSjob.GetJobType(/datum/job/xenomorph)
 	if(left_behind)
 		xeno_message("[left_behind > 1 ? "[left_behind] sisters" : "One sister"] perished due to being too slow to board the bird. The freeing of their psychic link allows us to call burrowed, at least.")
-		var/datum/job/xeno_job = SSjob.GetJobType(/datum/job/xenomorph)
 		xeno_job.add_job_positions(left_behind)
-		if(difference > 0)
-			xeno_job.add_job_positions(difference)
+	if(difference < 0)
+		if(xeno_job.total_positions < -difference)
+			xeno_job.set_job_positions(-difference)
 	for(var/i in GLOB.xeno_resin_silos)
+		if(isalamoarea(get_area(i)))
+			continue
+		if(!is_ground_level(i.z))
+			continue
 		qdel(i)
